requirements.txt
load-dotenv==0.1.0
psycopg2==2.9.7
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
